+++ CHANGELOG.md
@@ -1,3 +1,41 @@
+## 2.3.0
+
+### Identity V3:
+
+In this version, we’ve redesigned the way that user identification works. 
+Detailed docs about the new system are available [here](https://docs.revenuecat.com/v3.2/docs/user-ids).
+
+#### New methods
+- Introduces `logIn`, a new way of identifying users, which also returns whether a new user has been registered in the system. 
+`logIn` uses a new backend endpoint. 
+- Introduces `logOut`, a replacement for `reset`. 
+
+#### Deprecations
+- deprecates `createAlias` in favor of `logIn`.
+- deprecates `identify` in favor of `logIn`.
+- deprecates `reset` in favor of `logOut`.
+- deprecates `setAllowSharingStoreAccount` in favor of dashboard-side configuration.
+    https://github.com/RevenueCat/cordova-plugin-purchases/pull/80
+
+#### Other
+- Fix return type in `canMakePayments` documentation
+    https://github.com/RevenueCat/cordova-plugin-purchases/pull/80
+- Dependency security updates
+    https://github.com/RevenueCat/cordova-plugin-purchases/pull/98
+- Bumped purchases-ios to 3.12.2 [Changelog here](https://github.com/RevenueCat/purchases-ios/releases/3.12.2)
+- Bumped purchases-android to 4.3.0 [Changelog here](https://github.com/RevenueCat/purchases-android/releases/4.3.0)
+- Updated BillingClient to version [4.0.0](https://developer.android.com/google/play/billing/release-notes#4-0).
+     https://github.com/RevenueCat/purchases-android/commit/f6554bbf7376c3fd492f0bc67183a9f35889ae78
+
+## 2.2.0
+
+- Add canMakePayments method (https://github.com/RevenueCat/cordova-plugin-purchases/pull/96)
+- Update type for availablePackages (https://github.com/RevenueCat/cordova-plugin-purchases/pull/94)
+- Cleaned up return types for sync methods (https://github.com/RevenueCat/cordova-plugin-purchases/pull/82)
+- Bumped purchases-hybrid-common to 1.7.1 [Changelog here](https://github.com/RevenueCat/purchases-hybrid-common/releases/1.7.1)
+- Bumped purchases-ios to 3.11.1 [Changelog here](https://github.com/RevenueCat/purchases-ios/releases/3.11.1)
+- Bumped purchases-android to 4.2.1 [Changelog here](https://github.com/RevenueCat/purchases-hybrid-common/releases/4.2.1)
+
 ## 2.1.1
 
 - Bumped purchases-hybrid-common to 1.6.2 [Changelog here](https://github.com/RevenueCat/purchases-hybrid-common/releases/1.6.2)